Migrate TableBundleRenderer
