--- javascript/app-framework/src/components/editors/TableBundleEditor/BundleEditor.types.ts
@@ -1,9 +1,0 @@
-/*!
- * Copyright (c) Microsoft. All rights reserved.
- * Licensed under the MIT license. See LICENSE file in the project.
- */
-import type { TableBundle } from '@datashaper/workflow'
-
-export interface BundleEditorProps {
-	resource: TableBundle
-}
