+++ javascript/app-framework/src/components/editors/TableBundleEditor/TableBundleEditor.hooks.ts
