+++ javascript/app-framework/src/profiles/renderers/TableBundleRenderer.tsx
@@ -4,31 +4,30 @@
  */
 import { memo } from 'react'
 
-import type { Codebook } from '@datashaper/workflow'
+import type { TableBundle } from '@datashaper/workflow'
 import { CompoundEditor } from '../../components/editors/CompoundEditor/index.js'
-import { CodebookEditor } from '../../components/editors/CodebookEditor/index.js'
 import { ResourceSchemaEditor } from '../../components/editors/ResourceSchemaEditor/index.js'
+import { TableBundleEditor } from '../../components/editors/TableBundleEditor/index.js'
 import { useHelpOnMount } from '../../hooks/useHelpOnMount.js'
 import type { PluginComponentProps } from '../../types.js'
 
-export const CodebookRenderer: React.FC<PluginComponentProps<Codebook>> = memo(
-	function CodebookRenderer({ resource }) {
-		useHelpOnMount('resources.codebook.index')
+export const TableBundleRenderer: React.FC<PluginComponentProps<TableBundle>> =
+	memo(function TableBundleRenderer({ resource }) {
+		useHelpOnMount('resources.tablebundle.index')
 		return <CompoundEditor resource={resource} editors={editors} />
-	},
-)
+	})
 
 const editors = [
 	{
-		key: 'codebook-json',
-		title: 'View and edit codebook JSON config',
+		key: 'tablebundle-json',
+		title: 'View and edit table bundle JSON config',
 		iconName: 'Code',
 		renderer: ResourceSchemaEditor,
 	},
 	{
-		key: 'codebook-interactive',
-		title: 'View and edit interactive codebook options',
+		key: 'tablebundle-interactive',
+		title: 'View and edit interactive table',
 		iconName: 'PreviewLink',
-		renderer: CodebookEditor,
+		renderer: TableBundleEditor,
 	},
 ]
